# docs/specifications/protocol/http-api.openapi.yaml
openapi: 3.0.3
info:
  title: AWSM Coordination HTTP API
  version: "1"
  description: >-
    Canonical pre-release HTTPS adapter for the transport-independent AWSM synchronization
    protocol. Opaque payload bytes are never described semantically by this contract.
servers:
  - url: /
security:
  - accountBearer: []
paths:
...
  /api/vaults/{vaultId}/generation-candidates:
    parameters:
      - $ref: "#/components/parameters/VaultId"
      - $ref: "#/components/parameters/ProtocolVersion"
      - $ref: "#/components/parameters/RequestId"
    post:
      operationId: createGenerationCandidate
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGenerationCandidateRequest"
      responses:
        "201":
          description: Inactive Generation candidate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerationCandidate"
        default:
          $ref: "#/components/responses/Outcome"
...
  /api/vaults/{vaultId}/generation-candidates/{generationId}/activate:
    parameters:
      - $ref: "#/components/parameters/VaultId"
      - $ref: "#/components/parameters/GenerationId"
      - $ref: "#/components/parameters/ProtocolVersion"
      - $ref: "#/components/parameters/RequestId"
      - $ref: "#/components/parameters/IdempotencyKey"
    post:
      operationId: activateGenerationCandidate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivateGenerationRequest"
      responses:
        "200":
          description: Activated successor Generation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vault"
        default:
          $ref: "#/components/responses/Outcome"
...
components:
  securitySchemes:
    accountBearer:
      type: http
      scheme: bearer
  headers:
    ProtocolVersion:
      required: true
      schema:
        type: string
        enum: ["1"]
    RequestId:
      required: true
      schema:
        $ref: "#/components/schemas/Uuid"
  parameters:
    ProtocolVersion:
      in: header
      name: Awsm-Protocol-Version
      required: true
      schema:
        type: string
        enum: ["1"]
    RequestId:
      in: header
      name: Awsm-Request-ID
      required: true
      schema:
        $ref: "#/components/schemas/Uuid"
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      schema:
        $ref: "#/components/schemas/Uuid"
    VaultId:
      in: path
      name: vaultId
      required: true
      schema:
        $ref: "#/components/schemas/Uuid"
    GenerationId:
      in: path
      name: generationId
      required: true
      schema:
        $ref: "#/components/schemas/Uuid"
...
  responses:
    Outcome:
      description: Stable protocol outcome
      headers:
        Awsm-Protocol-Version:
          $ref: "#/components/headers/ProtocolVersion"
        Awsm-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Outcome"
  schemas:
...
    Uuid:
      type: string
      format: uuid
      pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
    Counter:
      type: integer
      minimum: 0
      maximum: 9007199254740991
    PositiveByteLength:
      type: integer
      minimum: 1
      maximum: 9007199254740991
    Sha256:
      type: string
      pattern: "^[A-Za-z0-9_-]{43}$"
    ObjectType:
      type: string
      enum: [Event, BundleDescriptor, Artifact, VaultGeneration]
    ObjectMetadata:
      type: object
      additionalProperties: false
      required: [objectId, objectType, keyEpochId, byteLength, sha256]
      properties:
        objectId:
          $ref: "#/components/schemas/Uuid"
        objectType:
          $ref: "#/components/schemas/ObjectType"
        keyEpochId:
          $ref: "#/components/schemas/Uuid"
        byteLength:
          $ref: "#/components/schemas/PositiveByteLength"
        sha256:
          $ref: "#/components/schemas/Sha256"
...
    Vault:
      type: object
      additionalProperties: false
      required:
        [
          vaultId,
          state,
          generationId,
          generationNumber,
          headCursor,
          activeKeyEpochId,
        ]
      properties:
        vaultId:
          $ref: "#/components/schemas/Uuid"
        state:
          type: string
          enum: [Provisional, Active, Replaced]
        generationId:
          $ref: "#/components/schemas/Uuid"
        generationNumber:
          $ref: "#/components/schemas/Counter"
        predecessorGenerationId:
          $ref: "#/components/schemas/Uuid"
        headCursor:
          $ref: "#/components/schemas/Counter"
        activeKeyEpochId:
          $ref: "#/components/schemas/Uuid"
...
    Upload:
      type: object
      additionalProperties: false
      required:
        [
          uploadId,
          objectId,
          state,
          partSizeBytes,
          partCount,
          receivedParts,
          expiresAt,
        ]
      properties:
        uploadId:
          $ref: "#/components/schemas/Uuid"
        objectId:
          $ref: "#/components/schemas/Uuid"
        state:
          type: string
          enum: [Open, Assembling, Completed, Expired, AlreadyDurable]
        partSizeBytes:
          $ref: "#/components/schemas/PositiveByteLength"
        partCount:
          type: integer
          minimum: 1
          maximum: 10000
        receivedParts:
          type: array
          uniqueItems: true
          items:
            type: integer
            minimum: 0
        expiresAt:
          type: string
          format: date-time
...
    TransferTicket:
      type: object
      additionalProperties: false
      required: [method, url, expiresAt, requiredHeaders]
      properties:
        method:
          type: string
          enum: [GET, PUT]
        url:
          type: string
        urlTemplate:
          type: string
        expiresAt:
          type: string
          format: date-time
        requiredHeaders:
          type: object
          additionalProperties:
            type: string
...
    CreateGenerationCandidateRequest:
      type: object
      additionalProperties: false
      required:
        [
          generationId,
          generationNumber,
          predecessorGenerationId,
          headCursor,
          generationObject,
        ]
      properties:
        generationId:
          $ref: "#/components/schemas/Uuid"
        generationNumber:
          $ref: "#/components/schemas/Counter"
        predecessorGenerationId:
          $ref: "#/components/schemas/Uuid"
        headCursor:
          $ref: "#/components/schemas/Counter"
        generationObject:
          $ref: "#/components/schemas/ObjectMetadata"
    GenerationCandidate:
      type: object
      additionalProperties: false
      required:
        [
          generationId,
          generationNumber,
          predecessorGenerationId,
          baselineCursor,
          state,
        ]
      properties:
        generationId:
          $ref: "#/components/schemas/Uuid"
        generationNumber:
          $ref: "#/components/schemas/Counter"
        predecessorGenerationId:
          $ref: "#/components/schemas/Uuid"
        baselineCursor:
          $ref: "#/components/schemas/Counter"
        state:
          type: string
          enum: [Candidate, Sealed]
        upload:
          $ref: "#/components/schemas/Upload"
        ticket:
          $ref: "#/components/schemas/TransferTicket"
...
    ActivateGenerationRequest:
      type: object
      additionalProperties: false
      required:
        [predecessorGenerationId, predecessorGenerationNumber, headCursor]
      properties:
        predecessorGenerationId:
          $ref: "#/components/schemas/Uuid"
        predecessorGenerationNumber:
          $ref: "#/components/schemas/Counter"
        headCursor:
          $ref: "#/components/schemas/Counter"
...
    Outcome:
      type: object
      additionalProperties: false
      required: [outcome, retryable, requestId]
      properties:
        outcome:
          type: string
          enum:
            - AUTHENTICATION_FAILED
            - AUTHENTICATION_UNAVAILABLE
            - ACCOUNT_UNAVAILABLE
            - ACCOUNT_INPUT_INVALID
            - RECENT_AUTHENTICATION_REQUIRED
            - AUTHORIZATION_FAILED
            - PROTOCOL_VERSION_UNSUPPORTED
            - REQUEST_INVALID
            - IDEMPOTENCY_CONFLICT
            - ACCOUNT_VAULT_LIMIT
            - VAULT_ID_UNAVAILABLE
            - VAULT_NOT_FOUND
            - VAULT_NOT_READY
            - RECOVERY_GENERATION_CHANGED
            - KEY_EPOCH_CHANGED
            - DEVICE_REVOKED
            - DEVICE_ENROLLMENT_INVALID
            - VAULT_REPLACEMENT_CONFLICT
            - OBJECT_ID_CONFLICT
            - OBJECT_NOT_DURABLE
            - OBJECT_NOT_ACTIVE
            - OBJECT_CHECKSUM_MISMATCH
            - OBJECT_LENGTH_MISMATCH
            - UPLOAD_EXPIRED
            - UPLOAD_PART_CONFLICT
            - TRANSFER_TICKET_INVALID
            - DEPENDENCY_INVALID
            - VAULT_GENERATION_SUPERSEDED
            - VAULT_HEAD_CHANGED
            - GENERATION_CANDIDATE_CONFLICT
            - GENERATION_REACHABILITY_INVALID
            - RECOVERY_NOT_FOUND
            - RECOVERY_EXPIRED
            - PURGE_IN_PROGRESS
            - SERVER_BUSY
            - STORAGE_UNAVAILABLE
        retryable:
          type: boolean
        requestId:
          $ref: "#/components/schemas/Uuid"
        retryAfterSeconds:
          type: integer
          minimum: 0
        relatedObjectId:
          $ref: "#/components/schemas/Uuid"
        currentGenerationId:
          $ref: "#/components/schemas/Uuid"
        currentGenerationNumber:
          $ref: "#/components/schemas/Counter"
        headCursor:
          $ref: "#/components/schemas/Counter"
...